refactor daemon executor plan ownership
